sim2real/pyproject.toml
[project]
name = "gmt-sim2real"
version = "0.1.0"
description = "Sim2Real and Sim2Sim runtime for G1 tracking"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
  "linuxfd",
  "mujoco",
  "numpy",
  "onnxruntime",
  "pin>=3.9.0",
  "pyyaml",
  "pyzmq>=27.1.0",
  "scipy",
  "sshkeyboard",
  "unitree-sdk2py",
]

[tool.uv.sources]
unitree-sdk2py = { git = "https://github.com/EGalahad/unitree_sdk2_python.git" }

[tool.uv]
# This repo is script-oriented; only manage deps/env with uv.
package = false
